refactor(storage): make discard own live download retirement

+++ lib/core/services/book_storage_service.dart
@@ -6,6 +6,7 @@ import 'package:night_reader/core/database/dao/reader_chapter_content_dao.dart';
 import 'package:night_reader/core/di/injection.dart';
 import 'package:night_reader/core/models/book.dart';
 import 'package:night_reader/core/services/book_cover_storage_service.dart';
+import 'package:night_reader/core/services/download_service.dart';
 
 class BookStorageService {
   BookStorageService({
@@ -14,6 +15,7 @@ class BookStorageService {
     ReaderChapterContentDao? contentDao,
     BookmarkDao? bookmarkDao,
     DownloadDao? downloadDao,
+    DownloadService? downloadService,
     BookCoverStorageService? coverStorage,
   }) : _bookDao = bookDao ?? getIt<BookDao>(),
        _chapterDao = chapterDao ?? getIt<ChapterDao>(),
@@ -28,6 +30,9 @@ class BookStorageService {
        _downloadDao =
            downloadDao ??
            (getIt.isRegistered<DownloadDao>() ? getIt<DownloadDao>() : null),
+       _downloadService =
+           downloadService ??
+           (getIt.isRegistered<DownloadDao>() ? DownloadService() : null),
        _coverStorage = coverStorage ?? BookCoverStorageService();
 
   final BookDao _bookDao;
@@ -35,12 +40,17 @@ class BookStorageService {
   final ReaderChapterContentDao? _contentDao;
   final BookmarkDao? _bookmarkDao;
   final DownloadDao? _downloadDao;
+  final DownloadService? _downloadService;
   final BookCoverStorageService _coverStorage;
 
   Future<void> discardBook(Book book) async {
+    if (_downloadService != null) {
+      await _downloadService!.retireBook(book.bookUrl);
+    } else {
+      await _downloadDao?.deleteByUrl(book.bookUrl);
+    }
     await _contentDao?.deleteByBook(book.origin, book.bookUrl);
     await _bookmarkDao?.deleteByBook(book.bookUrl);
-    await _downloadDao?.deleteByUrl(book.bookUrl);
     await _chapterDao.deleteByBook(book.bookUrl);
     await _bookDao.deleteByUrl(book.bookUrl);
     await _coverStorage.deleteBookAssets(book);
